+++ TennisClub.DAL/Configurations/MemberConfiguration.cs
@@ -17,7 +17,7 @@ namespace TennisClub.DAL.Configurations
             builder.Property(i => i.FirstName).HasColumnType("varchar(25)").IsRequired();
             builder.Property(i => i.LastName).HasColumnType("varchar(35)").IsRequired();
             builder.Property(i => i.BirthDate).HasColumnType("date").IsRequired();
-            builder.Property(i => i.GenderId).HasColumnType("tinyint").IsRequired();
+            builder.Property(i => i.GenderId).HasColumnType("integer").IsRequired();
             builder.Property(i => i.Address).HasColumnType("varchar(70)").IsRequired();
             builder.Property(i => i.Number).HasColumnType("varchar(6)").IsRequired();
             builder.Property(i => i.Addition).HasColumnType("varchar(2)").IsRequired(false);
