+++ TennisClub.DAL/Configurations/MemberRoleConfiguration.cs
@@ -14,7 +14,7 @@ namespace TennisClub.DAL.Configurations
 
             builder.Property(i => i.Id).HasColumnType("integer").UseIdentityColumn();
             builder.Property(i => i.MemberId).HasColumnType("integer").IsRequired();
-            builder.Property(i => i.RoleId).HasColumnType("tinyint").IsRequired();
+            builder.Property(i => i.RoleId).HasColumnType("integer").IsRequired();
             builder.Property(i => i.StartDate).HasColumnType("date").IsRequired();
             builder.Property(i => i.EndDate).HasColumnType("date"); // TODO: vragen in combinatie met de unique key. .IsRequired(false); // Is onmogelijk.
