+++ TennisClub.DAL/Entities/Game.cs
@@ -13,7 +13,7 @@ namespace TennisClub.DAL.Entities
         public int Id { get; set; }
         public string GameNumber { get; set; }
         public int MemberId { get; set; }
-        public byte LeagueId { get; set; }
+        public int LeagueId { get; set; }
         public DateTime Date { get; set; }
 
         public ICollection<GameResult> GameResults { get; set; }
